requirements.txt
aiobotocore
python-dotenv
